Separate ui and kit
