+++ devtools/ships/General.tsx
@@ -4,12 +4,7 @@ import { useFormContext } from "react-hook-form";
 import { ShipInput, shipRoles } from "@core/world/ships";
 import { Textures, textures } from "@core/components/render";
 import { DockSize } from "@core/components/dockable";
-import {
-  Select,
-  SelectButton,
-  SelectOption,
-  SelectOptions,
-} from "@ui/components/Select";
+import { Select, SelectButton, SelectOption, SelectOptions } from "@kit/Select";
 import { Table, TableCell, TableHeader } from "../components/Table";
 import { styles } from "./styles";
 import { FormData, useThrottledFormState } from "./utils";
