+++ ui/components/AutoOrder.tsx
@@ -4,9 +4,9 @@ import { Order } from "@core/components/orders";
 import { RequireComponent } from "@core/tsHelpers";
 import okIcon from "@assets/ui/ok.svg";
 import { nano, theme } from "../style";
-import { Select, SelectButton, SelectOption, SelectOptions } from "./Select";
-import { IconButton } from "./IconButton";
-import Text from "./Text";
+import { Select, SelectButton, SelectOption, SelectOptions } from "@kit/Select";
+import { IconButton } from "@kit/IconButton";
+import Text from "@kit/Text";
 import { useSim } from "../atoms";
 
 const styles = nano.sheet({
