+++ ui/components/ConfigDialog.tsx
@@ -4,11 +4,11 @@ import { Sim } from "@core/sim";
 import { Save } from "@core/db";
 import arrowLeftIcon from "@assets/ui/arrow_left.svg";
 import { nano, theme } from "../style";
-import { Dialog } from "./Dialog";
-import { Button } from "./Button";
-import { Input } from "./Input";
+import { Dialog } from "@kit/Dialog";
+import { Button } from "@kit/Button";
+import { Input } from "@kit/Input";
 import { useLocation } from "../context/Location";
-import { IconButton } from "../components/IconButton";
+import { IconButton } from "@kit/IconButton";
 import { Saves } from "./Saves";
 import useFullscreen from "../hooks/useFullscreen";
 import { useSim } from "../atoms";
