+++ ui/components/ContextMenu/ShipToEntity.tsx
@@ -5,7 +5,7 @@ import { isOwnedByPlayer } from "@core/components/player";
 import { getSelected, getSelectedSecondary } from "@core/components/selection";
 import { moveToOrders } from "@core/utils/moving";
 import { useContextMenu, useGameDialog, useSim } from "../../atoms";
-import { DropdownOption } from "../Dropdown";
+import { DropdownOption } from "@kit/Dropdown";
 import { NoAvailableActions } from "./NoAvailableActions";
 
 export const ShipToEntity: React.FC = () => {
