+++ ui/components/Docks.tsx
@@ -4,7 +4,7 @@ import {
   Collapsible,
   CollapsibleContent,
   CollapsibleSummary,
-} from "./Collapsible";
+} from "@kit/Collapsible";
 import { EntityList } from "./EntityList";
 
 export interface DocksProps {
