+++ ui/components/Journal.tsx
@@ -5,7 +5,7 @@ import { RequireComponent } from "@core/tsHelpers";
 import { limitMax, limitMin } from "@core/utils/limit";
 import arrowLeftIcon from "@assets/ui/arrow_left.svg";
 import { nano, theme } from "../style";
-import { IconButton } from "./IconButton";
+import { IconButton } from "@kit/IconButton";
 
 const styles = nano.sheet({
   arrowRight: {
