+++ ui/components/SectorStats.tsx
@@ -9,9 +9,9 @@ import { Sector } from "@core/archetypes/sector";
 import { getSectorResources } from "@core/utils/resources";
 import { fieldColors } from "@core/archetypes/asteroid";
 import { nano, theme } from "../style";
-import { Button } from "./Button";
-import { Dialog } from "./Dialog";
-import { Checkbox } from "./Checkbox";
+import { Button } from "@kit/Button";
+import { Dialog } from "@kit/Dialog";
+import { Checkbox } from "@kit/Checkbox";
 
 const styles = nano.sheet({
   commodities: {
