+++ ui/components/Subordinates.tsx
@@ -5,7 +5,7 @@ import {
   Collapsible,
   CollapsibleContent,
   CollapsibleSummary,
-} from "./Collapsible";
+} from "@kit/Collapsible";
 import { EntityList } from "./EntityList";
 
 export interface SubordinatesProps {
