+++ ui/views/NewGame.tsx
@@ -4,9 +4,9 @@ import { Sim } from "@core/sim";
 import world from "@core/world";
 import { nano, theme } from "../style";
 import { useLocation } from "../context/Location";
-import { Slider } from "../components/Slider";
-import { Button } from "../components/Button";
-import Text from "../components/Text";
+import { Slider } from "@kit/Slider";
+import { Button } from "@kit/Button";
+import Text from "@kit/Text";
 import { View } from "../components/View";
 import { useWorker } from "../hooks/useWorker";
 import { HeadlessSimMsg } from "../workers/headlessSim";
